+++ src/memory/Memory.java
@@ -3,21 +3,21 @@ package memory;
 import java.util.Hashtable;
 
 public class Memory {
-	Hashtable<Integer, String> memory;
+	Hashtable<Long, String> memory;
 	
 	public Memory() {
-		memory = new Hashtable<Integer, String>(); 
+		memory = new Hashtable<Long, String>(); 
 	}
 	
-	public String read(Integer address) {
+	public String read(Long address) {
 		return memory.get(address);
 	}
 	
-	public String write(Integer address, String contents) {
+	public String write(Long address, String contents) {
 		return memory.put(address, contents);
 	}
 	
-	public Hashtable<Integer, String> getAll() {
+	public Hashtable<Long, String> getAll() {
 		return memory;
 	}
